+++ src/page/index.ts
@@ -1,0 +1,13 @@
+import { Rule } from '@angular-devkit/schematics';
+import { uiComponent } from '../ui-component';
+
+interface Options {
+    name: string;
+}
+
+export function page(options: Options): Rule {
+    return uiComponent({
+        ...options,
+        type: 'page',
+    });
+}