+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
+## v0.2.0
+
+- feat(generator): support page section card and tile schematics
+
 ## v0.1.0
 
 - init(project): Automated project init
 - feat(generator): Declare first generators
+- chore(changelog): Publish v0.1.0